+++ app/src/main/kotlin/com/teakave/noteapp/presentation/feature/note/detail/DetailFragment.kt
@@ -1,10 +1,13 @@
 package com.teakave.noteapp.presentation.feature.note.detail
 
+import android.annotation.SuppressLint
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.activity.OnBackPressedCallback
+import androidx.core.view.isInvisible
+import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import com.teakave.domain.feature.note.model.NoteData
@@ -13,12 +16,16 @@ import com.teakave.noteapp.presentation.feature.common.KeyboardUtil
 import com.teakave.noteapp.presentation.feature.note.viewmodel.NoteViewModel
 import org.koin.android.ext.android.inject
 import org.koin.androidx.viewmodel.ext.android.sharedViewModel
+import java.text.SimpleDateFormat
 import java.util.Date
+
+const val LAST_MODIFIED_DATE_FORMAT = "dd.MMMM.yyyy HH:mm"
 
 class DetailFragment : Fragment() {
 
     private val viewModel by sharedViewModel<NoteViewModel>()
     private val keyboardUtil by inject<KeyboardUtil>()
+
     private var _binding: FragmentNoteDetailBinding? = null
     private val binding get() = _binding!!
 
@@ -28,6 +35,7 @@ class DetailFragment : Fragment() {
      * Initialization in [onViewCreated] function.
      */
     private var note: NoteData? = null
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -59,17 +67,28 @@ class DetailFragment : Fragment() {
             object : OnBackPressedCallback(true) {
                 override fun handleOnBackPressed() {
                     viewModel.setSelectedNote(null)
-                    saveNote()
-                    findNavController().navigateUp()
+                    if (contentChanged()) {
+                        saveNote()
+                    }
+                    navigateUp()
                 }
             })
 
     /**
-     * Update note title and content.
+     * Update note title, content and timestamp.
      */
+    @SuppressLint("SimpleDateFormat")
     private fun updateUI(noteData: NoteData?) = noteData?.let {
-        binding.editTextNoteTitle.setText(it.title)
-        binding.editTextNoteContent.setText(it.content)
+        binding.apply {
+            editTextNoteTitle.setText(it.title)
+            editTextNoteContent.setText(it.content)
+            textNoteLastUpdate.apply {
+                isVisible = true
+                text = SimpleDateFormat(LAST_MODIFIED_DATE_FORMAT).format(it.lastUpdateDate)
+            }
+        }
+    } ?: run {
+        binding.textNoteLastUpdate.isInvisible = true
     }
 
     /**
@@ -80,8 +99,13 @@ class DetailFragment : Fragment() {
      * This function calls saveNote function in our viewModel.
      */
     private fun setupOnSaveClickedListener() = binding.floatingButtonSaveNote.setOnClickListener {
-        setupSaveNoteResultObserver()
-        saveNote()
+        if (contentChanged()) {
+            setupSaveNoteResultObserver()
+            saveNote()
+        } else {
+            viewModel.setSelectedNote(null)
+            navigateUp()
+        }
     }
 
     /**
@@ -106,14 +130,24 @@ class DetailFragment : Fragment() {
         viewModel.saveNoteResult.observe(viewLifecycleOwner, androidx.lifecycle.Observer {
             if (it.isSuccess()) {
                 hideKeyboard()
-                findNavController().navigateUp()
+                navigateUp()
             } // todo: show a message if we can not save the note
         })
     }
 
     /**
+     * Determine if user changed the original content of the note.
+     *
+     * Comparing title and detail.
+     */
+    private fun contentChanged() = (binding.editTextNoteTitle.text.toString() != note?.title
+            || binding.editTextNoteContent.text.toString() != note?.content)
+
+    /**
      * Hides keyboard
      */
     private fun hideKeyboard() = keyboardUtil.hideKeyboard(view)
+
+    private fun navigateUp() = findNavController().navigateUp()
 
 }